requirements.txt
# API and Web Framework
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3

# External Service Integrations
slack_sdk>=3.23.0
requests>=2.31.0

# Database
sqlalchemy>=2.0.22
alembic>=1.12.0

# Scheduling
apscheduler>=3.10.4
pytz>=2023.3

# AI Integration
google-generativeai>=0.3.0
boto3>=1.28.57  # For AWS Bedrock

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
jinja2>=3.1.2
markdown>=3.5

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
httpx>=0.25.0
mongomock>=4.1.2